backend/requirements.txt
# FastAPI e servidor
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17

# Configurações
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.1

# Banco de dados
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.36
alembic>=1.14.0

# Redis
redis>=5.2.0

# OpenAI
openai>=1.59.0

# HTTP Client (para Meta API)
httpx>=0.28.0

# Autenticação
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0

# WebSocket
websockets>=14.0

# Utilitários
python-dateutil>=2.9.0
pytz>=2024.2

# Twilio
twilio>=9.0.0
